Reorganize components into subdirectories
