+++ frontend/src/components/environment/Camera.js
