+++ frontend/src/components/environment/Lights.js
