+++ frontend/src/components/mainView/Words.js
